started merging chat from paxxium, I have to account for chat settings

+++ app/routes/chat_route.py
@@ -24,38 +24,64 @@ def handle_socketio_options():
     response.headers['Access-Control-Allow-Methods'] = 'GET, POST, OPTIONS'
     return response
 
-@chat_bp.route('/chat', methods=['OPTIONS'])
+
 @chat_bp.route('/messages', methods=['OPTIONS'])
-def cors_preflight_response():
-    cors_headers = {
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "GET, POST, OPTIONS, DELETE, PUT, PATCH",
-        "Access-Control-Allow-Headers": "Authorization, Content-Type, Project-ID",
-        "Access-Control-Max-Age": "3600",
-    }
-    return ("", 204, cors_headers)
+def handle_messages_options():
+    if request.method == 'OPTIONS':
+        return ("", 204)
+@chat_bp.route('/chat', defaults={'subpath': ''}, methods=['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'])
+@chat_bp.route('/chat/<path:subpath>', methods=['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'])
+def chat(subpath):
+    if request.method == 'OPTIONS':
+        return ("", 204)
 
-@chat_bp.route('/chat', methods=['GET'])
-def handle_fetch_chats():
-    user_id = request.headers.get('userId')
-    return chat_service.get_all_chats(user_id), 200, headers
+    if request.method == 'GET' and subpath == '':
+        user_id = request.headers.get('userId')
+        return chat_service.get_all_chats(user_id), 200, headers
+    
+    if request.method == 'POST' and subpath == '':
+        data = request.get_json()
+        chat_id = chat_service.create_chat_in_db(
+        data['userId'], 
+        data['chatName'], 
+        data['model'], 
+        system_prompt=data.get('systemPrompt'), 
+        chat_constants=data.get('chatConstants'), 
+        use_profile_data=data.get('useProfileData')
+    )
+        return {
+            'chatId': chat_id,
+            'chat_name': data['chatName'],
+            'model': data['model'],
+            'userId': data['userId'],
+            'systemPrompt': data.get('systemPrompt'),
+            'chatConstants': data.get('chatConstants'),
+            'useProfileData': data.get('useProfileData')
+        }, 200, headers
 
-@chat_bp.route('/chat', methods=['POST'])
-def handle_create_chat():
-    data = request.get_json()
-    chat_id = chat_service.create_chat_in_db(data['userId'], data['chatName'], data['model'])
-    return {
-        'chatId': chat_id,
-        'chat_name': data['chatName'],
-        'model': data['model'],
-        'userId': data['userId'],
-    }, 200, headers
+    if request.method == 'DELETE' and subpath == '':
+        chat_id = request.get_json()['chatId']
+        chat_service.delete_chat(chat_id)
+        return 'Conversation deleted', 200, headers
+    
+    if subpath == 'update_visibility':
+        data = request.get_json()
+        chat_id = data['chatId']
+        is_open = data['is_open']
+        chat_service.update_visibility(chat_id, is_open)
+        return ('Chat visibility updated', 200, headers)
 
-@chat_bp.route('/chat', methods=['DELETE'])
-def handle_delete_chat():
-    chat_id = request.get_json()['chatId']
-    chat_service.delete_chat(chat_id)
-    return 'Conversation deleted', 200, headers
+    if subpath == 'update_settings':
+        data = request.get_json()
+        use_profile_data = data.get('use_profile_data')
+        chat_name = data.get('chat_name')
+        chat_id = data.get('chatId')
+        agent_model = data.get('agent_model')
+        system_prompt = data.get('system_prompt')
+        chat_constants = data.get('chat_constants')
+        chat_service.update_settings(chat_id, chat_name, agent_model, system_prompt, chat_constants, use_profile_data)
+
+        return ('Chat settings updated', 200, headers)
 
 @socketio.on('join_room')
 def handle_join_room(data):
